+++ src/Livewire/Products/ProductShow.php
@@ -2,19 +2,22 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\Products;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\Product;
 
 class ProductShow extends Component
 {
-    use Toast;
+    use AuthorizesRequests, Toast;
 
     public Product $product;
 
     public function delete($id)
     {
         if ($product = Product::find($id)) {
+            $this->authorize('delete', $product);
+
             $product->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.product_deleted')), redirectTo: route('laravel-crm.products.index'));
