+++ src/Livewire/Quotes/QuoteIndex.php
@@ -5,6 +5,7 @@ namespace VentureDrake\LaravelCrm\Livewire\Quotes;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
 use Livewire\Component;
@@ -19,7 +20,7 @@ use VentureDrake\LaravelCrm\Traits\ResetsPaginationWhenPropsChanges;
 
 class QuoteIndex extends Component
 {
-    use ClearsProperties, ResetsPaginationWhenPropsChanges, SearchesEncryptableContacts, Toast, WithPagination;
+    use AuthorizesRequests, ClearsProperties, ResetsPaginationWhenPropsChanges, SearchesEncryptableContacts, Toast, WithPagination;
 
     public $layout = 'index';
 
@@ -127,6 +128,8 @@ class QuoteIndex extends Component
     public function delete($id)
     {
         if ($quote = Quote::find($id)) {
+            $this->authorize('delete', $quote);
+
             $quote->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.quote_deleted')));
@@ -136,6 +139,8 @@ class QuoteIndex extends Component
     public function accept($id): void
     {
         if ($quote = Quote::find($id)) {
+            $this->authorize('update', $quote);
+
             $quote->update([
                 'accepted_at' => Carbon::now(),
                 'pipeline_stage_id' => $this->pipeline->pipelineStages()->where('name', 'Accepted')->first()->id ?? null,
@@ -148,6 +153,8 @@ class QuoteIndex extends Component
     public function reject($id): void
     {
         if ($quote = Quote::find($id)) {
+            $this->authorize('update', $quote);
+
             $quote->update([
                 'rejected_at' => Carbon::now(),
                 'pipeline_stage_id' => $this->pipeline->pipelineStages()->where('name', 'Rejected')->first()->id ?? null,
@@ -160,6 +167,8 @@ class QuoteIndex extends Component
     public function unaccept($id): void
     {
         if ($quote = Quote::find($id)) {
+            $this->authorize('update', $quote);
+
             $quote->update([
                 'accepted_at' => null,
                 'pipeline_stage_id' => $this->pipeline->pipelineStages()->where('name', 'Draft')->first()->id ?? null,
@@ -172,6 +181,8 @@ class QuoteIndex extends Component
     public function unreject($id): void
     {
         if ($quote = Quote::find($id)) {
+            $this->authorize('update', $quote);
+
             $quote->update([
                 'rejected_at' => null,
                 'pipeline_stage_id' => $this->pipeline->pipelineStages()->where('name', 'Draft')->first()->id ?? null,
